+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-api/src/main/java/org/apache/shardingsphere/data/pipeline/spi/importer/ImporterCreatorFactory.java
@@ -15,26 +15,30 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.data.pipeline.core.datasource;
+package org.apache.shardingsphere.data.pipeline.spi.importer;
 
-import org.apache.shardingsphere.data.pipeline.api.datasource.PipelineDataSourceWrapper;
-import org.apache.shardingsphere.data.pipeline.api.datasource.config.PipelineDataSourceConfiguration;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+import org.apache.shardingsphere.infra.util.spi.ShardingSphereServiceLoader;
+import org.apache.shardingsphere.infra.util.spi.type.typed.TypedSPIRegistry;
 
 /**
- * Pipeline data source manager.
+ * Importer creator factory.
  */
-public interface PipelineDataSourceManager {
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public final class ImporterCreatorFactory {
+    
+    static {
+        ShardingSphereServiceLoader.register(ImporterCreator.class);
+    }
     
     /**
-     * Get cached data source.
+     * Get importer creator instance.
      *
-     * @param dataSourceConfig data source configuration
-     * @return data source
+     * @param databaseType database type
+     * @return importer creator
      */
-    PipelineDataSourceWrapper getDataSource(PipelineDataSourceConfiguration dataSourceConfig);
-    
-    /**
-     * Close, close cached data source.
-     */
-    void close();
+    public static ImporterCreator getInstance(final String databaseType) {
+        return TypedSPIRegistry.getRegisteredService(ImporterCreator.class, databaseType);
+    }
 }
